[SHRINKDESC-92] All descriptors are now generated by the metadata-plugin
