+++ api-misc/.gitignore
@@ -1,0 +1,2 @@
+src/main/java
+
