+++ impl-misc/src/main/resources/META-INF/services/org.jboss.shrinkwrap.descriptor.api.jetty7.JettyDescriptor
@@ -1,0 +1,3 @@
+implClass=org.jboss.shrinkwrap.descriptor.impl.jetty7.JettyDescriptorImpl
+importerClass=org.jboss.shrinkwrap.descriptor.spi.node.dom.XmlDomNodeDescriptorImporterImpl
+defaultName=default.xml
