+++ metadata-parser/src/main/java/org/jboss/shrinkwrap/descriptor/metadata/MetadataParserPath.java
@@ -14,6 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.jboss.shrinkwrap.descriptor.metadata;
