+++ metadata-parser/src/main/java/org/jboss/shrinkwrap/descriptor/metadata/filter/ElementFilter.java
@@ -25,6 +25,35 @@ import org.w3c.dom.Element;
 import org.w3c.dom.Node;
 import org.w3c.dom.traversal.TreeWalker;
 
+
+/**
+ * This class analyzes <code>Element</code> w3c elements.
+ * <p>
+ * Definition:
+ * <p>
+ * <code>
+ *   <element
+ *      abstract = boolean : false
+ *      block = (#all | List of (extension | restriction | substitution))
+ *      default = string
+ *      final = (#all | List of (extension | restriction))
+ *      fixed = string
+ *      form = (qualified | unqualified)
+ *      id = ID
+ *      maxOccurs = (nonNegativeInteger | unbounded)  : 1
+ *      minOccurs = nonNegativeInteger : 1
+ *      name = NCName
+ *      nillable = boolean : false
+ *      ref = QName
+ *      substitutionGroup = QName
+ *      type = QName
+ *      {any attributes with non-schema namespace . . .}>
+ *      Content: (annotation?, ((simpleType | complexType)?, (unique | key | keyref)*))
+ *   </element>
+ * </code>
+ * 
+ * @author <a href="mailto:ralf.battenfeld@bluewin.ch">Ralf Battenfeld</a>
+ */
 public class ElementFilter implements Filter
 {
    @Override
