+++ metadata-parser/src/main/java/org/jboss/shrinkwrap/descriptor/metadata/filter/ExtensionFilter.java
@@ -25,6 +25,12 @@ import org.w3c.dom.Element;
 import org.w3c.dom.Node;
 import org.w3c.dom.traversal.TreeWalker;
 
+
+/**
+ * This class analyzes <code>Extension</code> w3c elements.
+ * 
+ * @author <a href="mailto:ralf.battenfeld@bluewin.ch">Ralf Battenfeld</a>
+ */
 public class ExtensionFilter implements Filter
 {
    public boolean filter(final Metadata metadata, final TreeWalker walker)
