+++ metadata-parser/src/main/java/org/jboss/shrinkwrap/descriptor/metadata/filter/RestrictionFilter.java
@@ -24,6 +24,12 @@ import org.w3c.dom.Element;
 import org.w3c.dom.Node;
 import org.w3c.dom.traversal.TreeWalker;
 
+
+/**
+ * Class which is responsible for <code>Restriction</code> w3c elements.
+ * 
+ * @author <a href="mailto:ralf.battenfeld@bluewin.ch">Ralf Battenfeld</a>
+ */
 public class RestrictionFilter implements Filter
 {
    public boolean filter(final Metadata metadata, final TreeWalker walker)
