+++ metadata-parser/src/main/resources/META-INF/ddJavaAll.xsl
@@ -1015,7 +1015,7 @@
     <!-- ****************************************************** -->
     <xsl:function name="xdd:CheckDataType" as="xs:string">
         <xsl:param name="pTypeName"/>
-        <xsl:message select="concat('xdd:CheckDataType: ', $pTypeName)"/>
+<!--         <xsl:message select="concat('xdd:CheckDataType: ', $pTypeName)"/> -->
         <xsl:choose>
             <xsl:when test=" starts-with($pTypeName, 'xsd:')">
                 <xsl:sequence select="xdd:getJavaDataType($pTypeName)"/>
